updated the timer and included some more options
The timer now actually switches to a long rest when it should, not the break after.
Updated the theme switching so that it now works with all themes, provided the theme itself is correctly defined.
added more config options to allow for greater veriability for the timers that run during operation.
increased the feedback given when switching between breaks and work times, including better tracking of the amount of breaks left before a long break as well as a reminder when you are on break which break you are taking.